requirements.txt
pillow>=9.0.0
piexif>=1.1.3